requirements.txt
flask
Flask-SQLAlchemy
flask-login
gunicorn==20.1.0
markdown
bleach
